requirements.txt
Flask>=3.0,<4.0
requests>=2.31,<3.0
openpyxl>=3.1,<4.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<24.0
